% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosnma.run.R
\name{crosnma.run}
\alias{crosnma.run}
\title{Run Generic NMA or NMR model for dichotomous outcomes}
\usage{
crosnma.run(
  model,
  n.adapt = 1000,
  n.burnin = floor(n.iter/2),
  n.iter,
  thin = 1,
  n.chains = 3,
  quiet = TRUE
)
}
\arguments{
\item{model}{A \code{crosnmaModel} object produced by running \code{crosnma.model}.}

\item{n.adapt}{Number of adaptations for the mcmc chains.}

\item{n.burnin}{Number of burnin iterations for the mcmc chains.}

\item{n.iter}{Number of iterations for the mcmc chains.}

\item{thin}{Thinning factor for the mcmc chains. Default is 1.}

\item{n.chains}{Number of mcmc chains. Default is 3.
See \code{\link{jags.model}} for more info.}
}
\value{
\code{crosnma.run} returns an object of class \code{crosnmaRun} which is a list containing the following components:

\code{samples} - The MCMC samples produced by running the BUGS model.

\code{model} - The \code{BUGSnetModel} object obtained from \code{nma.model} which was used to run \code{jags}.

\code{trt.key} - Treatments mapped to numbers, used to run BUGS code.
}
\description{
Takes jags code from an object produced by \code{crosnma.model} and runs model using \code{jags}.
}
\seealso{
\code{\link{crosnma.model}},\code{\link{jags.model}}
}
