% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosnma.plot.R
\name{plot.crosnma}
\alias{plot.crosnma}
\title{Generic plot function for crosnma.model object.}
\usage{
\method{plot}{crosnma}(x, ...)
}
\arguments{
\item{x}{The object generated by the gnma.run function.}

\item{...}{\dots}
}
\description{
Generic plot function for crosnma.model object.
}
