% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{prt.data}
\alias{prt.data}
\title{A simulated individual participant data.}
\format{
A data frame with 2950 rows and 8 variables:
\describe{
  \item{study}{numeric, study identifier}
  \item{outcome}{binray data, respond indicator, 0=respond and 1=no-respond}
  \item{trt}{character, indicating the assigned treatment to each participant}
  \item{design}{character, design of the study, either 'rct' or 'nrs'}
  \item{age}{numeric, age of the participant}
  \item{sex}{binray data, sex of the participant, 0=Female and 1=Male}
  \item{bias}{character, the risk of bias of the study, 'low', 'high','unclear'}
  \item{year}{numeric, the year published of the study}

}
}
\usage{
prt.data
}
\description{
A dataset containing 2950 participants who are treated in 4 different treatments: A, B, C and D.
The dataset includes 4 different studies. The outcome is binary. There are 8 attributes on indvidual level.
}
\keyword{datasets}
