% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosnma.summary.R
\name{summary.crosnma}
\alias{summary.crosnma}
\title{Generic summary function for crosnma.run object}
\usage{
\method{summary}{crosnma}(x, digits = 3, quantiles = c(0.025, 0.5, 0.975), expo = TRUE, ...)
}
\arguments{
\item{x}{The object generated by the crosnma.run function.}

\item{digits}{The number of significant digits printed. The default value is 3.}

\item{quantiles}{A numeric vector of probabilities with values in [0,1]. The default value is
intervals = c(0.025, 0.5, 0.975).}

\item{expo}{A logical indicating whether to expontiate the parameters of relative treatment effect and covariate effect}

\item{...}{\dots}
}
\description{
Generic summary function for crosnma.run object
}
