% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosnma.run.R
\name{crosnma.run}
\alias{crosnma.run}
\title{Run JAGS to synthesize cross-design evidence and cross-format data in NMA and NMR for dichotomous outcomes}
\usage{
crosnma.run(
  model,
  n.adapt = 1000,
  n.burnin = floor(n.iter/2),
  n.iter,
  thin = 1,
  n.chains = 2,
  quiet = TRUE
)
}
\arguments{
\item{model}{A \code{crosnmaModel} object produced by \code{crosnma.model}.}

\item{n.adapt}{Number of adaptations for the MCMC chains.}

\item{n.burnin}{Number of burnin iterations for the MCMC chains.}

\item{n.iter}{Number of iterations for the MCMC chains.}

\item{thin}{Number of thinning for the MCMC chains. Default is 1.}

\item{n.chains}{Number of MCMC chains. Default is 2.}

\item{quiet}{A logical. If TRUE, the warning message will not be displayed
See \code{\link{jags.model}} for more info.}
}
\value{
\code{crosnma.run} returns an object of class \code{crosrun} which is a list containing the following components:

\code{samples}  The MCMC samples produced by running the JAGS model.

\code{model}  The \code{crosnmaModel} object obtained from \code{crosnma.model} which was used to run \code{jags}.

\code{trt.key}  A table of the treatment names and their correspondence to integers used in the JAGS model.
}
\description{
Takes the JAGS model from an object produced by \code{crosnma.model} and runs it using \code{jags} package.
}
\examples{
# An example from participant-level data and study-level data.
# data
data(prt.data)
data(std.data)
 #=========================#
  # Create a jags model  #
 #=========================#
 # We conduct a network meta-analysis assuming a random effect model.
 # The data comes from randomised-controlled trials and non-randomised studies. They will be combined naively.
 # The data has 2 different formats: individual participant data (prt.data) and study-level data (std.data).
mod <- crosnma.model(prt.data=prt.data,
                  std.data=std.data,
                  trt=c('trt','trt'),
                  study=c('study','study'),
                  outcome=c('outcome','outcome'),
                  n='n',
                  design=c('design','design'),
                  reference='A',
                  trt.effect='random',
                  covariate = NULL,
                  method.bias='naive'
                   )
 #=========================#
    # Fit jags model  #
 #=========================#
fit <- crosnma.run(model=mod,
                n.adapt = 20,
                n.iter=50,
                thin=1,
                n.chains=3)

 #=========================#
   # Display the output   #
 #=========================#
summary(fit)
plot(fit)


}
\seealso{
\code{\link{crosnma.model}},\code{\link{jags.model}}
}
